Added a bunch of comments about Minesweeper conventions (since I seem to be the "expert" in this group
Also, it seems that one of the constructors invoked in the button Class is "deprecated". Might be worthwhile to look at that later.
Also it seems that git is automattically changing my files from LF to CRLF????
IDK what if anything that means

Note to DCP: Got Maven up and running on my Windows PC, I'll take another stab at getting it to work on my Linux machine tomorrow.

Signed-off-by: IkeKap <1906011@wpsstudent.com>